feat(wip): Write logs to file
The charm's logging scraper expects logs to be written to a file.
The location of the file needs to also match the location specified in
the charm. Ideally the location of this file would be provided as an env
variable and then there would be no need to coordinated between the
charm and ubuntu-report.

So, this is a wip commit given that I don't have time to do this
properly, but I want to show the system wired up correctly.